backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
pydantic[email]==2.10.4
passlib[bcrypt]==1.7.4
PyJWT==2.12.1
bcrypt==4.0.1
httpx==0.28.1
pytest==8.3.4
python-dotenv==1.2.2
